NLCD land cover by percent of county area for each county in MO odd numbered years 2001-2023.

A dataset derived of Annual NLCD data- files were obtained via the FedData package (NLCD) and tigir package (county boundaries). NLCD data was reprojected to NAD83 UTM 15 N before areal calculations.

nlcd.pct.tab

Format

cty

county name

yr

4 digit year

crop.pct

percent area of crop

develop.pct

percent area of developed

forest.pct

percent area of forest

grass.pct

percent area of grass

water.pct

percent area of water

wetland.pct

percent area of wetland

other.pct

percent area of all other landcover types present

Details

Cropping performed with terra::crop

Area converted to square km based on a 30x30m cell size and summing all all pixels with in a given class. No other processing occurred.

Examples

if (FALSE) { # \dontrun{
#pixel classes were condensed as follows:

terra::rast(nlcd)->Z
 mo.co.list[[county]]->Q
 terra::crop(Z,Q)%>%
 terra::freq()%>%
 filter(str_detect(value,
 'Water|Developed|Forest|Shrub|Grass|Hay|Crops|Wetlands'))%>%
 mutate(LC=case_when(str_detect(value,'Developed')~"develop",
                   str_detect(value,'Forest')~"forest",
                   str_detect(value,'Shrub|Grassland|Hay')~'grass',
                   str_detect(value,'Crops')~'crop',
                   str_detect(value,'Wetlands')~'wetland',
                   str_detect(value,'Water')~'water',
                   TRUE~as.character("other")
                   ))
                   } # }