A simple survival simulation.

Simulate known or apparent survival data via a HMM, with or without interval censoring or right censoring. The base simulation simulates known-fate data, without interval or right censoring.

Usage

surv_sim1(
  S.int = 0.95,
  nind = 10,
  ntime = 10,
  p.obs = 0.5,
  last.obs = FALSE,
  obs.start = TRUE,
  censor.parms = c(0, 1),
  visit.parms = c(1, 1),
  type = "RAW"
)

Arguments

S.int

the interval survival probability

nind

the number of individuals

ntime

the length of the encounter histories

p.obs

detection probability for imperfect detection

last.obs

TRUE or FALSE. Is the terminal time interval observed. Typically TRUE in nest survival studies.

obs.start

TRUE or FALSE, if true, all individuals are observed at time =1, otherwise. all individuals may not be observed until later in their encounter history, or not at all before failing. They are not removed from the data object and must be excluded before use in analysis.

censor.parms

a vector of length 2, giving the probability an individual is censored and the probability defining the time an individual is censored. The time of censoring is generated by a binomial where the maximum length of the encounter history is the number of trials and the second parameter is the probability. The default values prevent any individual from being censored. Here, censoring is a known event (i.e. a known transmitter failure, drop, etc.)

type

LOGEXP, KF, CJS or SURV. CJS= A wide matrix of 1,0; apparent survival. KF= a wide matrix of 1(live), 0(unobserved), and 2(dead); MS HMM known-fate LOGEXP= a long data frame formatted for logistic exposure; known fate SURV= a long data frame of entry time, last observation, and if the observation was censored or an event was observed-suitable for survival analysis using the survival program.

Value

data

Details

Interval censoring (for logistic exposure or other known-fate designs) is accomplished by including an observation probability. This will inherently lead to right-censoring, but here, the observation at the last time interval can be fixed to 1 and interval censoring is simulated alternatively. Thus, when there is no formal right censoring, all dead individuals are observed dead at least one time.

Right censoring is created by providing two probabilities. The first defines the probability any individual will be censored. The second is used in a binomial probability where n_trials is the length of the observation period. While true right censoring can occur at any time, this better simulates more common causes of right censoring, like transmitters failing at the end of battery life.

Known-fate data requires individuals are observable in a dead state. When individuals cannot be observed in a dead state and detection is imperfect, apparent survival is the result. The same parameter that generates intervals for known-fate data with perfect detection can be re-purposed as detection probability for an apparent survival simulation with imperfect detection. Individuals that die are no longer observable as dead, and instead are treated as unobservable.